main,refactor: split read.h into read.h and read_p.h
