+++ main/debug.c
@@ -23,6 +23,7 @@
 #include "options.h"
 #include "parse_p.h"
 #include "read.h"
+#include "read_p.h"
 
 /*
 *   FUNCTION DEFINITIONS
