+++ main/entry.c
@@ -48,6 +48,7 @@
 #include "options_p.h"
 #include "ptag_p.h"
 #include "read.h"
+#include "read_p.h"
 #include "routines.h"
 #include "routines_p.h"
 #include "parse_p.h"
