+++ main/lxpath.c
@@ -14,6 +14,7 @@
 #include "options.h"
 #include "parse_p.h"
 #include "read.h"
+#include "read_p.h"
 #include "routines.h"
 #include "xtag.h"
