+++ main/main.c
@@ -70,7 +70,7 @@
 #include "main.h"
 #include "options_p.h"
 #include "parse_p.h"
-#include "read.h"
+#include "read_p.h"
 #include "routines_p.h"
 #include "trace.h"
 #include "trashbox.h"
